@@ -34,20 +34,20 @@ repository: https://github.com/ianmaciel/ahgora_api
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 0.1.8
+version: 0.3.0-nullsafety.0
 
 environment:
-  sdk: ">=2.6.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
-  http: ^0.12.0+4
-  intl: ^0.16.1
-  json_annotation: ^3.0.1
+  http: ^0.13.0
+  intl: ^0.17.0
+  json_annotation: ^4.0.0
 
 dev_dependencies:
   test:
-  build_runner: ^1.7.4
-  json_serializable: ^3.2.5
+  build_runner: ^1.11.5
+  json_serializable: ^4.0.2
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
