@@ -8,9 +8,7 @@ part of 'monthly_report.dart';
 
 MonthlyReport _$MonthlyReportFromJson(Map<String, dynamic> json) {
   return MonthlyReport(
-    json['funcionario'] == null
-        ? null
-        : Employee.fromJson(json['funcionario'] as Map<String, dynamic>),
+    Employee.fromJson(json['funcionario'] as Map<String, dynamic>),
     MonthlyReport._referenceFromJson(json['referencia'] as String),
     MonthlyReport._daysFromJson(json['dias'] as Map<String, dynamic>),
   );
@@ -19,6 +17,6 @@ MonthlyReport _$MonthlyReportFromJson(Map<String, dynamic> json) {
 Map<String, dynamic> _$MonthlyReportToJson(MonthlyReport instance) =>
     <String, dynamic>{
       'funcionario': instance.employee,
-      'referencia': instance.reference?.toIso8601String(),
+      'referencia': instance.reference.toIso8601String(),
       'dias': instance.days,
     };
